Engine support C++20 now

+++ src/xrEngine/editor_environment_levels_manager.cpp
@@ -119,7 +119,7 @@ LPCSTR const* manager::collection	()
 
 u32 manager::collection_size		()
 {
-	return							(m_weathers.weather_ids().size());
+	return							u32(m_weathers.weather_ids().size());
 }
 
 void manager::fill					()
